Add files via upload
This is RytmTool a Max patch that allows MIDI manipulation of Analog Rytm. It adds additional 10 modulation sources and that can be assigned to any parameter of any voice. 
It can be used with other synths using NPRN MIDI just NPRN text file with appropriate data needs to be loaded. CC MIDI is also possible with slight modification.